refactor: typeorm comment entity 파일 이름 변경
