+++ src/modules/comments/comments.service.ts
@@ -6,7 +6,7 @@ import {
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { CommentsRepository } from './comments-typeorm.repository';
-import { Comment } from './entities/comment.entity';
+import { Comment } from './entities/comment-typeorm.entity';
 import { UsersTypeormRepository } from '../users/users-typeorm.repository';
 import { PostsRepository } from '../posts/posts-typeorm.repository';
 import { Post } from '../posts/entities/post.entity';
